Adds async/await to first methods

+++ server/server.js
@@ -67,27 +67,24 @@ app.get('/todos/:id', authenticate, (req, res) => {
         });
 });
 
-app.delete('/todos/:id', authenticate, (req, res) => {
+// esta
+app.delete('/todos/:id', authenticate, async (req, res) => {
     const id = req.params.id;
 
     if (!ObjectID.isValid(id)) {
         return res.status(404).send();
     }
 
-    Todo.findOneAndRemove({
-        _id: id,
-        _creator: req.user._id
-    })
-        .then(todo => {
-            if (!todo) {
-                return res.status(404).send();
-            }
+    try {
+        const todo = await Todo.findOneAndRemove({ _id: id, _creator: req.user._id });
+        if (!todo) {
+            return res.status(404).send();
+        }
 
-            res.send({ todo });
-        })
-        .catch(err => {
-            res.status(400).send();
-        });
+        res.send({ todo });
+    } catch (error) {
+        res.status(400).send();
+    }
 });
 
 app.patch('/todos/:id', authenticate, (req, res) => {
@@ -128,48 +125,41 @@ app.patch('/todos/:id', authenticate, (req, res) => {
         });
 });
 
-app.post('/users', (req, res) => {
+app.post('/users', async (req, res) => {
     const body = _.pick(req.body, ['email', 'password']);
     const user = new User(body);
 
-    user.save()
-        .then(() => {
-            return user.generateAuthToken();
-        })
-        .then(token => {
-            res.header('x-auth', token).send(user);
-        })
-        .catch(err => {
-            res.status(400).send(err);
-        });
+    try {
+        await user.save();
+        const token = await user.generateAuthToken();
+        res.header('x-auth', token).send(user);
+    } catch (error) {
+        res.status(400).send(error);
+    }
 });
 
 app.get('/users/me', authenticate, (req, res) => {
     res.send(req.user);
 });
 
-app.post('/users/login', (req, res) => {
-    const body = _.pick(req.body, ['email', 'password']);
-
-    User.findByCredentials(body.email, body.password)
-        .then(user => {
-            return user.generateAuthToken().then(token => {
-                res.header('x-auth', token).send(user);
-            });
-        })
-        .catch(err => {
-            res.status(400).send();
-        });
+app.post('/users/login', async (req, res) => {
+    try {
+        const body = _.pick(req.body, ['email', 'password']);
+        const user = await User.findByCredentials(body.email, body.password);
+        const token = await user.generateAuthToken()
+        res.header('x-auth', token).send(user);
+    } catch (error) {
+        res.status(400).send();
+    }
 });
 
-app.delete('/users/me/token', authenticate, (req, res) => {
-    req.user.removeToken(req.token)
-        .then(() => {
-            res.status(200).send();
-        })
-        .catch(() => {
-            res.status(400).send();
-        })
+app.delete('/users/me/token', authenticate, async (req, res) => {
+    try {
+        await req.user.removeToken(req.token)
+        res.status(200).send();
+    } catch (error) {
+        res.status(400).send();
+    }
 })
 
 app.listen(port, () => {
